+++ src/Ext/Techno/Hooks.cpp
@@ -1,8 +1,10 @@
 #include <InfantryClass.h>
+#include <ScenarioClass.h>
 
 #include "Body.h"
 
 #include <Ext/TechnoType/Body.h>
+#include <Ext/WarheadType/Body.h>
 #include <Ext/WeaponType/Body.h>
 
 DEFINE_HOOK(0x6F9E50, TechnoClass_AI, 0x5)
@@ -115,4 +117,38 @@ DEFINE_HOOK(0x6F3C88, TechnoClass_Transform_6F3AD0_BurstFLH_2, 0x6)
 		R->EAX(0);
 
 	return 0;
+}
+
+// Issue #237 NotHuman additional animations support
+// Author: Otamaa
+DEFINE_HOOK(0x518505, InfantryClass_TakeDamage_NotHuman, 0x4)
+{
+	GET(InfantryClass* const, pThis, ESI);
+	REF_STACK(args_ReceiveDamage const, receiveDamageArgs, STACK_OFFS(0xD0, -0x4));
+
+	// Die1-Die5 sequences are offset by 10
+	#define Die(x) x + 10
+
+	int resultSequence = Die(1);
+	auto const pTypeExt = TechnoTypeExt::ExtMap.Find(pThis->GetTechnoType());
+
+	if (pTypeExt->NotHuman_RandomDeathSequence.Get())
+		resultSequence = ScenarioClass::Instance->Random.RandomRanged(Die(1), Die(5));
+
+	if (receiveDamageArgs.WH)
+	{
+		if (auto const pWarheadExt = WarheadTypeExt::ExtMap.Find(receiveDamageArgs.WH))
+		{
+			int whSequence = pWarheadExt->NotHuman_DeathSequence.Get();
+			if (whSequence > 0)
+				resultSequence = Math::min(Die(whSequence), Die(5));
+		}
+	}
+
+	#undef Die(x)
+
+	R->ECX(pThis);
+	pThis->PlayAnim(static_cast<Sequence>(resultSequence), true);
+
+	return 0x518515;
 }